Db populate (#46)
* attack

* ACTot

* aC

* att

* attack

* bab

* feats

* enum

* show

* addFeat

* skill

* hp

* vitaError

* move

* sizeAr

* virgola

* sizeAc

* dto

* errorFeats

* attSize

* szSk

* grapple

* grapple

* specialAttack

* checkFeat

* featsList

* dupli

* featcheck

* featcheck

* feat

* switch

* feat

* rage49

* dupl

* dupli_DTO_method

* prerequisiteDTO

* sawintThrow_check

* sawintThrow_check_1

* Prerequisite_checkingi

* Prerequisite_checkingiV2

* empty_prer

* prerequisite_json

* characterF_dupli_classVSbuy

* add_featFromClassVSbuy

* checking_pre

* checking_preV2

* checking_preV3

* weapon

* weapon_dto

* weapon_dtoV1

* weapons_enum_in_DB

* weap

* weapons_get

* abstract?

* abstract

* req_weapon_main

* strategy inh

* Update WeaponsRepository

* Update idC1_weapons.xml

* weapons_not

* armor_error

* error

* hib_error

* findAll_Weapons

* list_items

* remove_item_object

* index nie działa

* index nie działa_1

* error_in_array

* ne ma itema :(

* sell_weapon_in_array

* arraylist of items

* DB update to fighter

* add monk and paladin

* int to short

* rouge DB

* skillsDB

* error_content-type

* constructor

* no UTF

* add and remove all skillsStudy is working

* fild of study

* charFeat lepiej

* correction DB

+++ api/src/main/java/pl/kolendateam/dadcard/classCharacter/entity/ClassPc.java
@@ -14,10 +14,10 @@ import lombok.Setter;
 @NoArgsConstructor
 public class ClassPc implements Serializable {
 
-    int id;
+    short id;
     String name;
-    int level;
-    int hitDice;
+    byte level;
+    byte hitDice;
     String savingThrow;
     double classBab;
 
@@ -37,7 +37,7 @@ public class ClassPc implements Serializable {
         return -1;
     }
 
-    public int findLevelInArrayById(ArrayList<ClassPc> classPcList, int id) {
+    public int findLevelInArrayById(ArrayList<ClassPc> classPcList, short id) {
 
         int levelOfClass=0;
