+++ api/src/main/java/pl/kolendateam/dadcard/race/entity/Region.java
@@ -16,7 +16,7 @@ import lombok.Setter;
 public class Region {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    int id;
+    short id;
 
     @Nonnull
     String name;
